Simplified video API
